+++ server/out/types/index.d.ts
@@ -1,5 +1,12 @@
 interface IReferences {
-    [key: string]: any;
+    [key: string]: IComponent;
+}
+interface IComponent {
+    properties: Record<string, string | Record<string, string>>;
+    methods: Record<string, string | string[]>;
+    signals: Record<string, string>;
+    inherit?: string;
+    doc?: string;
 }
 declare const References: IReferences;
 export default References;